pyproject.toml
[tool.poetry]
name = "nonebot_plugin_petpet"
version = "0.1.19"
description = "Nonebot2 plugin for making fun pictures"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-alpha.15"
nonebot-adapter-cqhttp = "^2.0.0-alpha.15"
aiocache = ">=0.11.0"
httpx = ">=0.19.0"
imageio = "^2.12.0"
numpy = "^1.20.0"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]
nb-cli = "^0.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
